+++ CHANGELOG.md
@@ -6,7 +6,8 @@
   Use lsblk instead of blkid if available.
 * [**Phil Dibowitz**:](https://github.com/jaymzh)
   linux::filesystem now reads all of /proc/mounts instead of just 4K
-
+* [**Phil Dibowitz**](https://github.com/jaymzh):
+  linux::network should handle ECMP routes
 
 ## Release: 7.2.0